WebSockets/tests/approved/interfaces/WebSocket/url/004.html

<!doctype html>
<title>WebSockets: 'URL'</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function() {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  assert_equals(ws.URL, undefined);
  assert_equals('URL' in ws, false);
  assert_equals(WebSocket.prototype.URL, undefined);
  assert_equals('URL' in WebSocket.prototype, false);
});
</script>
